[PATCH] Add library matching package name
[emacsmirror] This patch is being rebased on the Emacsmirror.
This package should be named `majapahit-themes' because it provides
more than one theme.  But upstream calls it `majapahit-theme' and
Melpa and Emacsmirror will continue to use that name as well, until
it is fixed upstream.  The file "majapahit-theme.el" only so we can
extract metadata from the (currently) expected place.

Original commit message, submitted to upstream:

Add library matching package name

Every package should provide a library whose name matches that of the
package.  This allows tools, such as those used by Melpa, to extract
metadata.

At the same time every library named "*-theme.el" must itself
implement a theme because `load-theme' unfortunately assumes that
every file on `custom-theme-load-path' with such a name actually
provides a theme.  If it encounters files that to not provide a theme
but are never-the-less named "*-theme", then it falsely offers those
as themes that can be activated.

Furthermore, since this package provides multiple themes, it makes
sense to give it a name that suggests so.

Unfortunately package.el does not (yet?) support informing the user
about package renames or even automatically updating to the new name,
so some users will not notice this rename for a while.  Add a note
at the beginning of the readme, in the hope that will help at least
some users.